Se mapea servicio de geolocalizacion con defectos en la misma debido a temas de geolocalizador, falta implementarlo en modales para ver la info de los equipos, se completa la app de escritorio al 90% aun sin instalador debido a que esta en pruebas, se decide cambiar la propiedad de longitud y latitud en la base a float y en la logica a double.

+++ EquipGo/EquipGo.Agent/Models/EquipoSyncRequestDto.cs
@@ -8,8 +8,8 @@
         public string MacEquipo { get; set; }
         public string SistemaOperativo { get; set; }
         public string VersionSoftware { get; set; }
-        public decimal Latitud { get; set; }
-        public decimal Longitud { get; set; }
+        public double Latitud { get; set; }
+        public double Longitud { get; set; }
 
         public string? CodigoBarras { get; set; }
         public int? IdUsuarioInfo { get; set; }
