+++ EquipGo/OUT_OS_APP.EQUIPGO.DTO/DTOs/Equipo/CrearEquipoDto.cs
@@ -18,8 +18,8 @@ namespace OUT_OS_APP.EQUIPGO.DTO.DTOs.Equipo
         public int? IdEquipoPersonal { get; set; }
         public int? IdSede { get; set; }
         public int? IdTipoDispositivo { get; set; }
-        public decimal? Latitud { get; set; }
-        public decimal? Longitud { get; set; }
+        public double? Latitud { get; set; }
+        public double? Longitud { get; set; }
         public string? SistemaOperativo { get; set; }
         public string? MacEquipo { get; set; }
         public string? VersionSoftware { get; set; }
